factor out passage text parts filtering
